requirements.txt
flask==0.10.1
flask-MongoKit==0.6
flask-restful==0.2.12
flask-oauth==0.12
flask-redis==0.0.6
python-dateutil==2.3

